feat(detectors): detect pod id on Kubernetes
This change introduces a new detector based on the [1]
upstream PR, to detect the Pod identifier of the Kubernetes
pod hosting the container.

As the pod identifier comes from /proc filesystems that
are used by other contaienr orchestrations, the detector
ensures that the container runs in Kubernetes by checking
whether the /etc/hosts file looks like it's managed by
Kubernetes.